+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 <!-- version list -->
 
+## v1.13.0 (2026-06-15)
+
+### Features
+
+- Add lockfile support to pywrangler sync command
+  ([#108](https://github.com/cloudflare/workers-py/pull/108),
+  [`4844aea`](https://github.com/cloudflare/workers-py/commit/4844aea44005052054b2d5ff5022b3dcd5ab4c49))
+
+
+## v1.12.0 (2026-06-12)
+
+### Features
+
+- Implements cf accessor on Request
+  ([`5777f80`](https://github.com/cloudflare/workers-py/commit/5777f80ead8d9a3c452fe3b6b8f2dc041d6c80d3))
+
+
 ## v1.11.0 (2026-06-03)
 
 ### Features